+++ selenium-testing/README.md
@@ -12,7 +12,7 @@ Instructions on how to install and run selenium tests
 
 ## Run tests
 
-First, copy and rename the file `config.properties.template` to `config.properties`, then fill in the missing details with base64 string encoded values
+First, copy and name the file `config.properties.template` to `config.properties`, then fill in the missing details with base64 string encoded values
 
 **`Step 1`** Add any test class tag names to `src/test/resources/testNG.xml` that you would like to run
