--- selenium-testing/src/main/java/com/testng/selenium/v1/App.java
@@ -1,13 +1,0 @@
-package com.testng.selenium.v1;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
