+++ selenium-testing/src/test/java/ca/bc/gov/ecc/ofm/selenium/v1/PageFactory_Portal/PortalApplicationsDeclareSubmit.java
@@ -1,4 +1,4 @@
-package com.testng.selenium.v1.PageFactory_Portal;
+package ca.bc.gov.ecc.ofm.selenium.v1.PageFactory_Portal;
 
 import java.time.Duration;
 import java.util.List;
@@ -18,28 +18,40 @@ import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class PortalApplicationsReview {
+public class PortalApplicationsDeclareSubmit {
 	private WebDriver driver;
 
-	
-
-	@FindBy(id = "app-next-button")
+	@FindBy(id = "declaration")
 	@CacheLookup
-	private WebElement next;
+	private WebElement checkboxDeclare;
+	
+	@FindBy(id = "app-submit-button")
+	@CacheLookup
+	private WebElement submit;
 
-	public PortalApplicationsReview(WebDriver driver) {
+
+	public PortalApplicationsDeclareSubmit(WebDriver driver) {
 
 		this.driver = driver;
 		PageFactory.initElements(driver, this);
 
 	}
 
-
-	public void clickNextButton() {
-		WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(20));
-		wait.until(ExpectedConditions.elementToBeClickable(next));
-		wait.until(ExpectedConditions.visibilityOf(next));
-		next.click();
+	public void clickDeclare() {
+		if (!checkboxDeclare.isSelected()) {
+			checkboxDeclare.click();
+		}
 
 	}
+	
+	public void clickSubmitButton() {
+		WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(20));
+		wait.until(ExpectedConditions.elementToBeClickable(submit));
+		wait.until(ExpectedConditions.visibilityOf(submit));
+		submit.click();
+
+	}
+	
+	
+	
 }
