+++ selenium-testing/src/test/java/ca/bc/gov/ecc/ofm/selenium/v1/tests/CRMLogin.java
@@ -1,4 +1,4 @@
-package com.testng.selenium.v1;
+package ca.bc.gov.ecc.ofm.selenium.v1.tests;
 
 import java.time.Duration;
 import org.openqa.selenium.WebDriver;
@@ -7,8 +7,7 @@ import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-import com.testng.selenium.v1.PageFactory_CRM.CRMSignInCredentialPage;
-
+import ca.bc.gov.ecc.ofm.selenium.v1.PageFactory_CRM.CRMSignInCredentialPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 
 public class CRMLogin extends BaseTest {
