+++ selenium-testing/src/test/java/ca/bc/gov/ecc/ofm/selenium/v1/tests/TestListener.java
@@ -1,4 +1,4 @@
-package com.testng.selenium.v1;
+package ca.bc.gov.ecc.ofm.selenium.v1.tests;
 
 import java.io.File;
 import java.io.IOException;
